+++ packages/core/src/components/inputs/CodeInput/codeInputValidation.ts
@@ -1,0 +1,6 @@
+export const allowedCodeCharacters = 'A-Z\\d'
+
+export const codeValidation = (value: string): boolean => {
+  const regex = `[${allowedCodeCharacters}]{8}$`
+  return value.match(regex) != null
+}