+++ packages/core/src/components/inputs/index.ts
@@ -3,6 +3,7 @@ import BaseInputV2 from '@core/components/inputs/BaseInputV2/BaseInputV2.vue'
 import BaseInputWrapper from '@core/components/inputs/BaseInputWrapper/BaseInputWrapper.vue'
 import BaseSelect from '@core/components/inputs/BaseSelect/BaseSelect.vue'
 import CheckboxInput from '@core/components/inputs/CheckboxInput/CheckboxInput.vue'
+import CurrencyInput from '@core/components/inputs/CurrencyInput/CurrencyInput.vue'
 import DateTimeInput from '@core/components/inputs/DateTimeInput/DateTimeInput.vue'
 import ImageInputWrapper from '@core/components/inputs/ImpageInputWrapper/ImageInputWrapper.vue'
 import LocationCoordinateSelect from '@core/components/inputs/LocationCoordinateSelect/LocationCoordinateSelect.vue'
@@ -11,6 +12,7 @@ import PreviewImageInput from '@core/components/inputs/PreviewImageInput/Preview
 import SelectInput from '@core/components/inputs/SelectInput/SelectInput.vue'
 import SliderToggleInput from '@core/components/inputs/SliderToggleInput/SliderToggleInput.vue'
 import UploadImageFAB from '@core/components/inputs/UploadImageFAB/UploadImageFAB.vue'
+import CodeInput from '@core/components/inputs/CodeInput/CodeInput.vue'
 
 export {
   BaseInput,
@@ -18,6 +20,8 @@ export {
   BaseInputWrapper,
   BaseSelect,
   CheckboxInput,
+  CurrencyInput,
+  CodeInput,
   DateTimeInput,
   ImageInputWrapper,
   LocationCoordinateSelect,