+++ packages/gridstack/src/components/ConfigurableDashboard/ConfigurableDashboard.vue
@@ -8,6 +8,7 @@
       @start-save="prepareSave"
       @cancel-edit="onCancelEdit"
       @add-widget="onAddWidget"
+      @delete-dashboard-configuration="onPrepareDelete"
       class="dashboard-bar"
     />
     <DynamicGrid
@@ -15,14 +16,31 @@
       v-model:widget-configs="dashboardConfig"
       :edit-mode="editMode"
     />
-    <DismissibleModal v-model:modalDisplayed="showModal" class="basic-modal">
-      <SaveDashboardModal v-if="showModal" @confirm-save="onConfirmSave" />
+    <DismissibleModal
+      v-model:modalDisplayed="showSaveModal"
+      class="basic-modal"
+    >
+      <SaveDashboardModal
+        v-if="showSaveModal"
+        :error="saveModalError"
+        @confirm-save="onConfirmSave"
+      />
+    </DismissibleModal>
+    <DismissibleModal
+      v-model:modal-displayed="showDeleteModal"
+      class="basic-modal"
+    >
+      <DeleteDashboardModal
+        v-if="showDeleteModal"
+        :configuration-name="selectedDashboardConfigurationName"
+        @confirm-delete="onDeleteDashboardConfiguration"
+      />
     </DismissibleModal>
   </div>
 </template>
 
 <script lang="ts" setup>
-import { nextTick, ref, watch } from 'vue'
+import { computed, nextTick, ref, watch } from 'vue'
 import { WidgetComponentWrapper } from '@/models/widgetComponentWrapper'
 import { WidgetConfiguration } from '@/models/widgetConfiguration'
 import { GridWidget } from '@/models/gridWidget'
@@ -30,8 +48,9 @@ import DashboardBar from '@/components/DashboardBar/DashboardBar.vue'
 import DynamicGrid from '@/components/DynamicGrid/DynamicGrid.vue'
 import { DismissibleModal } from '@lion5/component-library'
 import SaveDashboardModal from '@/components/SaveDashboardModal/SaveDashboardModal.vue'
+import DeleteDashboardModal from '@/components/DeleteDashboardModal/DeleteDashboardModal.vue'
 
-defineProps<{
+const props = defineProps<{
   /**
    * A map of widgets that can be added to the dashboard
    */
@@ -42,8 +61,26 @@ defineProps<{
   dashboardConfigurationOptions: Array<{ id: string; name: string }>
 }>()
 const emit = defineEmits<{
+  (e: 'deleteDashboardConfiguration', id: string): void
   (e: 'save', dashboardConfig: WidgetConfiguration[], name: string): void
 }>()
+
+/**
+ * A flag that indicates if the dashboard is in edit mode
+ */
+const editMode = defineModel<boolean>('editMode', {
+  default: false,
+  type: Boolean
+})
+
+/**
+ * A flag that indicates if save modal should be displayed
+ */
+const showSaveModal = defineModel<boolean>('showSaveModal')
+/**
+ * A flag that indicates if an error should be displayed in the save modal
+ */
+const saveModalError = defineModel<Error | undefined>('saveModalError')
 /**
  * Selected dashboard configuration
  */
@@ -62,23 +99,42 @@ const dashboardConfig = defineModel<WidgetConfiguration[]>('dashboardConfig', {
  * Currently set dashboardConfig
  */
 const currentConfig = ref(dashboardConfig.value)
-const editMode = ref<boolean>(false)
-const showModal = ref(false)
+const showDeleteModal = ref<boolean>(false)
+const selectedDashboardConfigurationName = computed(() => {
+  const config = props.dashboardConfigurationOptions.find(
+    (option) => option.id === selectedDashboardConfiguration.value
+  )
+  return config ? config.name : undefined
+})
 
 const prepareSave = () => {
-  showModal.value = true
+  showSaveModal.value = true
 }
 
 const onConfirmSave = (name: string) => {
-  editMode.value = false
   currentConfig.value = dashboardConfig.value
   emit('save', dashboardConfig.value, name)
-  showModal.value = false
 }
 
 const onCancelEdit = () => {
   editMode.value = false
   dashboardConfig.value = currentConfig.value
+  showSaveModal.value = false
+  saveModalError.value = undefined
+}
+
+const onDeleteDashboardConfiguration = () => {
+  if (!selectedDashboardConfiguration.value) {
+    return
+  }
+  emit('deleteDashboardConfiguration', selectedDashboardConfiguration.value)
+  selectedDashboardConfiguration.value = undefined
+  editMode.value = false
+  showDeleteModal.value = false
+}
+
+const onPrepareDelete = () => {
+  showDeleteModal.value = true
 }
 
 const onAddWidget = async (widgetKey: string) => {
@@ -101,6 +157,12 @@ const onAddWidget = async (widgetKey: string) => {
 watch(editMode, (isEditMode) => {
   if (isEditMode) {
     currentConfig.value = dashboardConfig.value
+  }
+})
+
+watch(showSaveModal, (isSaveModalDisplayed) => {
+  if (!isSaveModalDisplayed) {
+    saveModalError.value = undefined
   }
 })
 </script>